// lib/types.ts
export interface NavLink {
  href: string;
  label: string;
}

export interface Feature {
  image: string;
  title: string;
}

export interface TeamMember {
  name: string;
  role: string;
  image: string;
  link: string;
  details: string;
  department: string;
}

export interface Testimonial {
  text: string;
  author: string;
  avatar: string;
}

export interface FAQ {
  question: string;
  answer: string;
}

export interface SocialLink {
  href: string;
  label: string;
  icon: string;
}

export interface FooterLink {
  href: string;
  label: string;
}